requirements.txt
tensorflow==2.16.1
numpy
pillow
streamlit
matplotlib
